refactor: remove unused controller
