+++ graphene-reader/src/main/kotlin/com/graphene/reader/service/index/ElasticsearchClient.kt
@@ -2,12 +2,16 @@ package com.graphene.reader.service.index
 
 import net.iponweb.disthene.reader.config.IndexConfiguration
 import net.iponweb.disthene.reader.exceptions.TooMuchDataExpectedException
+import org.elasticsearch.action.search.SearchResponse
 import org.elasticsearch.action.search.SearchType
 import org.elasticsearch.client.transport.TransportClient
 import org.elasticsearch.common.unit.TimeValue
+import org.elasticsearch.index.query.FilteredQueryBuilder
 import org.elasticsearch.index.query.QueryBuilders
 import org.elasticsearch.search.SearchHits
+import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
+import javax.annotation.PreDestroy
 
 @Component
 class ElasticsearchClient(
@@ -33,39 +37,74 @@ class ElasticsearchClient(
 
     // if total hits exceeds maximum - abort right away returning empty array
     if (response.hits.totalHits() > indexConfiguration.maxPaths) {
-      ElasticsearchIndexService.logger.debug("Total number of paths exceeds the limit: " + response.hits.totalHits())
+      ElasticsearchIndexService.logger.debug("Total number map paths exceeds the limit: " + response.hits.totalHits())
       throw TooMuchDataExpectedException(
-        "Total number of paths exceeds the limit: "
+        "Total number map paths exceeds the limit: "
           + response.hits.totalHits()
           + " (the limit is "
           + indexConfiguration.maxPaths
           + ")")
     }
 
-    return Response(
-      response.scrollId,
-      response.hits,
-      response.hits.hits.size
-    )
+    return Response.of(response)
+  }
+
+  fun query(filteredQuery: FilteredQueryBuilder): Response {
+    var response = client
+      .prepareSearch(indexConfiguration.index)
+      .setScroll(TimeValue(indexConfiguration.timeout.toLong()))
+      .setSize(indexConfiguration.scroll)
+      .setQuery(filteredQuery)
+      .execute()
+      .actionGet()
+
+    // if total hits exceeds maximum - abort right away returning empty array
+    if (response.hits.totalHits() > indexConfiguration.maxPaths) {
+      ElasticsearchIndexService.logger.debug("Total number map paths exceeds the limit: " + response.hits.totalHits())
+      throw TooMuchDataExpectedException(
+        "Total number map paths exceeds the limit: "
+          + response.hits.totalHits()
+          + " (the limit is "
+          + indexConfiguration.maxPaths
+          + ")")
+    }
+
+    return Response.of(response)
   }
 
   fun searchScroll(response: Response): Response {
     val response = client.prepareSearchScroll(response.scrollId)
-      .setScroll(TimeValue.timeValueMinutes(1))
+      .setScroll(TimeValue.timeValueMinutes(indexConfiguration.timeout.toLong()))
       .execute()
       .actionGet()
 
-    return Response(
-      response.scrollId,
-      response.hits,
-      response.hits.hits.size
-    )
+    return Response.of(response)
   }
 
-  data class Response(
+  @PreDestroy
+  fun shutdown() {
+    logger.info("Shutting down elasticsearch client")
+    client.close()
+  }
+
+  data class Response internal constructor(
     val scrollId: String,
     val hits: SearchHits,
     val size: Int
-  )
+  ) {
+    companion object {
+      fun of(response: SearchResponse): Response {
+        return Response(
+          response.scrollId,
+          response.hits,
+          response.hits.hits.size
+        )
+      }
+    }
+  }
+
+  companion object {
+    internal val logger = LoggerFactory.getLogger(ElasticsearchClient::class.java)
+  }
 
 }
